requirements.txt
python-dotenv
langchain
langchain-groq
langchain_community
dash